renamed config to test-config. changed jest.config
